refactor: split star_schema.rs and streaming_builder.rs into modules
Split two large files in logfwd-arrow into module directories:
- star_schema.rs (3,615 lines) -> 5 files (helpers, flat_to_star, star_to_flat, tests)
- streaming_builder.rs (2,800 lines) -> 4 files (finish, tests, verification)

Pure structural refactor — no logic changes. All 232 tests pass.
