+++ crates/logfwd-arrow/src/streaming_builder/tests.rs
@@ -1,1226 +1,13 @@
-// streaming_builder.rs -- Zero-copy hot-path builder.
-//
-// Uses Arrow's StringViewArray to store string values as 16-byte views
-// pointing directly into the input buffer. No string copies during scanning
-// or during finish_batch/finish_batch_detached. The input buffer (as bytes::Bytes) is reference-counted
-// and shared between the builder and the resulting Arrow arrays.
-//
-// For the hot pipeline path: scan -> query -> output -> discard.
-
-use std::collections::HashMap;
-use std::sync::Arc;
-
-use arrow::array::{
-    ArrayRef, BooleanArray, Float64Array, Int64Array, StringViewBuilder, StructArray,
-};
-use arrow::buffer::{Buffer, NullBuffer};
-use arrow::datatypes::{DataType, Field, Fields, Schema};
-use arrow::error::ArrowError;
-use arrow::record_batch::{RecordBatch, RecordBatchOptions};
-use logfwd_lint_attrs::hot_path;
-
-use logfwd_core::scan_config::{parse_float_fast, parse_int_fast};
-use logfwd_core::scanner::BuilderState;
-
-use logfwd_types::field_names;
-
-use crate::check_dup_bits;
-use crate::columnar::row_protocol::RowLifecycle;
-
-#[cfg(kani)]
-type FieldIndexMap = std::collections::BTreeMap<Vec<u8>, usize>;
-#[cfg(not(kani))]
-type FieldIndexMap = HashMap<Vec<u8>, usize>;
-
-#[cfg(kani)]
-type EmittedNameSet = std::collections::BTreeSet<String>;
-#[cfg(not(kani))]
-type EmittedNameSet = std::collections::HashSet<String>;
-
-#[cfg(kani)]
-fn new_field_index() -> FieldIndexMap {
-    std::collections::BTreeMap::new()
-}
-
-#[cfg(not(kani))]
-fn new_field_index() -> FieldIndexMap {
-    HashMap::with_capacity(32)
-}
-
-fn new_emitted_name_set() -> EmittedNameSet {
-    EmittedNameSet::new()
-}
-
-fn append_string_view(
-    builder: &mut StringViewBuilder,
-    original_block: u32,
-    decoded_block: Option<u32>,
-    original_len: u32,
-    offset: u32,
-    len: u32,
-) -> Result<(), ArrowError> {
-    if offset < original_len || (len == 0 && offset == original_len) {
-        builder.try_append_view(original_block, offset, len)
-    } else if let Some(decoded_block) = decoded_block {
-        let decoded_offset = offset.checked_sub(original_len).ok_or_else(|| {
-            ArrowError::InvalidArgumentError("decoded string offset underflow".to_string())
-        })?;
-        builder.try_append_view(decoded_block, decoded_offset, len)
-    } else {
-        Err(ArrowError::InvalidArgumentError(format!(
-            "string view offset {offset} requires decoded buffer, but decoded buffer is absent"
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Per-field state
-// ---------------------------------------------------------------------------
-
-struct FieldColumns {
-    name: Vec<u8>,
-    /// String values: (row, offset_in_buffer, len). Views into the shared
-    /// buffer. Offsets `< buf.len()` reference the original input buffer;
-    /// offsets `>= buf.len()` reference the decoded-strings buffer at
-    /// `offset - buf.len()`. See `StreamingBuilder::decoded_buf`.
-    str_views: Vec<(u32, u32, u32)>,
-    /// Int values: (row, parsed_value).
-    int_values: Vec<(u32, i64)>,
-    /// Float values: (row, parsed_value).
-    float_values: Vec<(u32, f64)>,
-    /// Bool values: (row, value).
-    bool_values: Vec<(u32, bool)>,
-    has_str: bool,
-    has_int: bool,
-    has_float: bool,
-    has_bool: bool,
-    /// The last row this field was written to, used for dedup when idx >= 64.
-    last_row: u32,
-}
-
-impl FieldColumns {
-    fn new(name: &[u8]) -> Self {
-        FieldColumns {
-            name: name.to_vec(),
-            str_views: Vec::with_capacity(256),
-            int_values: Vec::with_capacity(256),
-            float_values: Vec::with_capacity(256),
-            bool_values: Vec::with_capacity(256),
-            has_str: false,
-            has_int: false,
-            has_float: false,
-            has_bool: false,
-            last_row: u32::MAX,
-        }
-    }
-
-    fn clear(&mut self) {
-        self.str_views.clear();
-        self.int_values.clear();
-        self.float_values.clear();
-        self.bool_values.clear();
-        self.has_str = false;
-        self.has_int = false;
-        self.has_float = false;
-        self.has_bool = false;
-        self.last_row = u32::MAX;
-    }
-}
-
-// ---------------------------------------------------------------------------
-// StreamingBuilder
-// ---------------------------------------------------------------------------
-
-/// Zero-copy builder for the hot pipeline path.
-///
-/// String values are stored as (offset, len) views into the input buffer.
-/// The buffer is reference-counted via `bytes::Bytes` and shared with the
-/// resulting Arrow `StringViewArray` columns -- no string copies at any point.
-///
-/// Numeric values (int, float) are parsed during scanning and stored directly.
-///
-/// When `line_field_name` is set, that column is emitted in each batch as a
-/// zero-copy `StringViewArray` containing the full unparsed line per row.
-///
-/// # Usage
-/// ```ignore
-/// let mut builder = StreamingBuilder::new(None);
-/// builder.begin_batch(bytes::Bytes::from(buf));
-/// // ... scan fields, call append_*_by_idx ...
-/// let batch = builder.finish_batch();
-/// ```
-pub struct StreamingBuilder {
-    fields: Vec<FieldColumns>,
-    field_index: FieldIndexMap,
-    /// Number of fields active in the current batch. Slots `0..num_active` in
-    /// `fields` are in use; slots beyond that are pre-allocated but dormant.
-    /// Resetting this to zero on `begin_batch` — together with clearing
-    /// `field_index` — bounds `fields.len()` to the high-water mark of unique
-    /// fields seen in any *single* batch rather than growing without limit.
-    num_active: usize,
-    /// Row lifecycle state machine: batch/row phase, row count, dedup bits.
-    lifecycle: RowLifecycle,
-    /// Reference-counted buffer. Stored here to compute offsets safely
-    /// and shared with Arrow StringViewArrays in finish_batch.
-    buf: bytes::Bytes,
-    /// Secondary buffer for decoded string values (JSON escape sequences).
-    /// String views with offsets `>= buf.len()` reference this buffer at
-    /// `offset - buf.len()`. Allocated lazily; empty when no escapes are decoded.
-    decoded_buf: Vec<u8>,
-    /// Optional output column name used for full-line capture.
-    line_field_name: Option<String>,
-    /// Line views: (offset_in_buf, len) per row, in row order.
-    /// Populated only when `line_field_name` is set.
-    line_views: Vec<(u32, u32)>,
-    /// Constant per-row resource attributes emitted as resource attribute columns.
-    resource_attrs: Vec<(String, String)>,
-}
-
-impl Default for StreamingBuilder {
-    fn default() -> Self {
-        Self::new(None)
-    }
-}
-
-impl StreamingBuilder {
-    pub fn new(line_field_name: Option<String>) -> Self {
-        StreamingBuilder {
-            fields: Vec::with_capacity(32),
-            field_index: new_field_index(),
-            num_active: 0,
-            lifecycle: RowLifecycle::new(),
-            buf: bytes::Bytes::new(),
-            decoded_buf: Vec::new(),
-            line_field_name,
-            line_views: Vec::new(),
-            resource_attrs: Vec::new(),
-        }
-    }
-
-    /// Configure constant per-row resource attributes for subsequent batches.
-    pub fn set_resource_attributes(&mut self, attrs: &[(String, String)]) {
-        self.resource_attrs.clear();
-        self.resource_attrs.extend(attrs.iter().cloned());
-    }
-
-    fn resource_col_name(key: &str) -> String {
-        // Canonical prefix + verbatim key. No mangling.
-        format!("{}{key}", field_names::DEFAULT_RESOURCE_PREFIX)
-    }
-
-    /// Start a new batch. Takes ownership of the input buffer via Bytes
-    /// (zero-copy if converted from Vec via `Bytes::from(vec)`).
-    ///
-    /// # Panics
-    /// Asserts that `buf.len()` fits in a u32, because string-view
-    /// offsets are stored as u32. Buffers larger than 4 GiB would produce
-    /// silently truncated offsets without this guard.
-    pub fn begin_batch(&mut self, buf: bytes::Bytes) {
-        debug_assert!(
-            matches!(
-                self.lifecycle.state(),
-                BuilderState::Idle | BuilderState::InBatch
-            ),
-            "begin_batch called while inside a row (missing end_row)"
-        );
-        assert!(
-            u32::try_from(buf.len()).is_ok(),
-            "StreamingBuilder buffer too large for u32 offsets ({} bytes)",
-            buf.len()
-        );
-        self.buf = buf;
-        self.decoded_buf.clear();
-        self.lifecycle.begin_batch();
-        // Only clear the slots that were active in the previous batch.
-        // This preserves the inner-Vec capacity of each FieldColumns for
-        // hot-path reuse while still bounding memory under key churn.
-        for fc in &mut self.fields[..self.num_active] {
-            fc.clear();
-        }
-        // Discard all name→index mappings so resolve_field rebuilds them from
-        // scratch.  Combined with resetting num_active, this prevents the
-        // HashMap and the fields Vec from growing across batches with churning
-        // field names (issue: field_index HashMap grows unboundedly).
-        self.field_index.clear();
-        self.num_active = 0;
-        self.line_views.clear();
-    }
-
-    #[hot_path]
-    #[inline(always)]
-    pub fn begin_row(&mut self) {
-        self.lifecycle.begin_row();
-    }
-
-    #[hot_path]
-    #[inline(always)]
-    pub fn end_row(&mut self) {
-        self.lifecycle.end_row();
-    }
-
-    #[inline]
-    pub fn resolve_field(&mut self, key: &[u8]) -> usize {
-        debug_assert!(
-            self.lifecycle.state() == BuilderState::InBatch
-                || self.lifecycle.state() == BuilderState::InRow,
-            "resolve_field called outside of an active batch"
-        );
-        // COLUMN NAME NOTE: The raw JSON key bytes are used verbatim as the
-        // output Arrow column name.  ScanConfig::is_wanted matches field names
-        // case-insensitively, so if a single input batch contains both `level`
-        // and `Level` they will pass the wanted filter independently and each
-        // call to resolve_field will allocate a *separate* column (because the
-        // HashMap lookup here is case-sensitive).  This means the two variants
-        // end up in distinct columns rather than being merged.  This is a known
-        // limitation: callers that need deterministic column naming should
-        // normalise JSON key casing before scanning, or post-process the batch.
-        if let Some(&idx) = self.field_index.get(key) {
-            return idx;
-        }
-        let idx = self.num_active;
-        if idx < self.fields.len() {
-            // Reuse an existing slot — its data was cleared in begin_batch.
-            // Only the name needs updating.
-            self.fields[idx].name.clear();
-            self.fields[idx].name.extend_from_slice(key);
-        } else {
-            self.fields.push(FieldColumns::new(key));
-        }
-        self.num_active += 1;
-        self.field_index.insert(key.to_vec(), idx);
-
-        idx
-    }
-
-    /// Compute the byte offset of `value` within `self.buf`.
-    ///
-    /// `value` must be a subslice of `self.buf`; i.e. both its pointer and
-    /// length lie within `self.buf` bounds. We use `usize` subtraction rather
-    /// than `offset_from` to avoid that API's stricter UB preconditions.
-    ///
-    /// # Panics
-    /// Panics if `value` does not lie entirely within `self.buf`, or if the
-    /// computed byte offset exceeds `u32::MAX` (buffer larger than 4 GiB).
-    /// These checks fire in both debug and release builds to prevent silent
-    /// StringView data corruption.
-    #[inline(always)]
-    fn offset_of(&self, value: &[u8]) -> u32 {
-        let base = self.buf.as_ptr() as usize;
-        let ptr = value.as_ptr() as usize;
-        assert!(
-            ptr >= base && ptr + value.len() <= base + self.buf.len(),
-            "value must be within buffer bounds"
-        );
-        let offset = ptr - base;
-        u32::try_from(offset).unwrap_or_else(|_| {
-            panic!("StreamingBuilder buffer offset exceeds u32::MAX ({offset} bytes)")
-        })
-    }
-
-    #[inline(always)]
-    pub fn append_str_by_idx(&mut self, idx: usize, value: &[u8]) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_str_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        if idx >= u64::BITS as usize && self.fields[idx].last_row == self.lifecycle.row_count() {
-            return;
-        }
-        // StringViewArray requires valid UTF-8.  JSON is always UTF-8 in
-        // production; for fuzz / corrupted input we skip non-UTF-8 bytes so
-        // that append_view_unchecked is never called on invalid bytes.
-        if std::str::from_utf8(value).is_err() {
-            return;
-        }
-        let Ok(len) = u32::try_from(value.len()) else {
-            return;
-        };
-        if idx >= u64::BITS as usize {
-            self.fields[idx].last_row = self.lifecycle.row_count();
-        }
-        let offset = self.offset_of(value);
-        let fc = &mut self.fields[idx];
-        fc.has_str = true;
-        fc.str_views.push((self.lifecycle.row_count(), offset, len));
-    }
-
-    #[inline(always)]
-    pub fn append_validated_str_by_idx(&mut self, idx: usize, value: &[u8]) {
-        self.append_str_by_idx(idx, value);
-    }
-
-    /// Append a decoded string value that is NOT a subslice of the input
-    /// buffer. Used for strings whose JSON escape sequences have been decoded
-    /// (see issue #410). Appends the bytes to `decoded_buf` and records a
-    /// view in the same `str_views` vector as regular strings, with the
-    /// offset shifted by `buf.len()` so that `finish_batch` can select the
-    /// decoded Arrow StringView block without copying the original input.
-    #[inline(always)]
-    pub fn append_decoded_str_by_idx(&mut self, idx: usize, value: &[u8]) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_decoded_str_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        if idx >= u64::BITS as usize && self.fields[idx].last_row == self.lifecycle.row_count() {
-            return;
-        }
-        if std::str::from_utf8(value).is_err() {
-            return;
-        }
-        self.append_decoded_str_inner(idx, value);
-    }
-
-    /// Append a pre-validated string (e.g. a prost `String` or `&str`) as a
-    /// decoded field value without re-running the UTF-8 validation check.
-    ///
-    /// **Caller contract**: `value` must be valid UTF-8. Violating this produces
-    /// corrupt Arrow arrays (undefined behaviour in downstream consumers).
-    #[inline(always)]
-    pub fn append_prevalidated_str_by_idx(&mut self, idx: usize, value: &str) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_prevalidated_str_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        if idx >= u64::BITS as usize && self.fields[idx].last_row == self.lifecycle.row_count() {
-            return;
-        }
-        self.append_decoded_str_inner(idx, value.as_bytes());
-    }
-
-    /// Inner append path shared by `append_decoded_str_by_idx` (post-validation)
-    /// and `append_prevalidated_str_by_idx` (type-guaranteed UTF-8).
-    #[inline(always)]
-    fn append_decoded_str_inner(&mut self, idx: usize, value: &[u8]) {
-        let Ok(len) = u32::try_from(value.len()) else {
-            return;
-        };
-        // Compute both offsets before mutating decoded_buf so that a bail-out
-        // on overflow does not leave unreferenced bytes in decoded_buf.
-        let Ok(decoded_offset) = u32::try_from(self.decoded_buf.len()) else {
-            // decoded_buf has grown past 4 GiB; drop this field rather than panic.
-            return;
-        };
-        // Offset into the combined buffer: original buf bytes come first,
-        // decoded bytes follow at buf.len() + decoded_offset.
-        let Some(combined_offset) = u32::try_from(self.buf.len())
-            .ok()
-            .and_then(|buf_len| buf_len.checked_add(decoded_offset))
-        else {
-            // Combined offset would overflow u32; drop this field rather than panic.
-            return;
-        };
-        // All validation passed — safe to update dedup guard and extend decoded_buf.
-        if idx >= u64::BITS as usize {
-            self.fields[idx].last_row = self.lifecycle.row_count();
-        }
-        self.decoded_buf.extend_from_slice(value);
-        let fc = &mut self.fields[idx];
-        fc.has_str = true;
-        fc.str_views
-            .push((self.lifecycle.row_count(), combined_offset, len));
-    }
-
-    #[inline(always)]
-    pub fn append_validated_decoded_str_by_idx(&mut self, idx: usize, value: &[u8]) {
-        self.append_decoded_str_by_idx(idx, value);
-    }
-
-    #[inline(always)]
-    pub fn append_int_by_idx(&mut self, idx: usize, value: &[u8]) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_int_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        let fc = &mut self.fields[idx];
-        if idx >= 64 {
-            if fc.last_row == self.lifecycle.row_count() {
-                return;
-            }
-            fc.last_row = self.lifecycle.row_count();
-        }
-        if let Some(v) = parse_int_fast(value) {
-            fc.has_int = true;
-            fc.int_values.push((self.lifecycle.row_count(), v));
-        }
-    }
-
-    #[inline(always)]
-    pub fn append_i64_value_by_idx(&mut self, idx: usize, value: i64) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_i64_value_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        let fc = &mut self.fields[idx];
-        if idx >= 64 {
-            if fc.last_row == self.lifecycle.row_count() {
-                return;
-            }
-            fc.last_row = self.lifecycle.row_count();
-        }
-        fc.has_int = true;
-        fc.int_values.push((self.lifecycle.row_count(), value));
-    }
-
-    #[inline(always)]
-    pub fn append_float_by_idx(&mut self, idx: usize, value: &[u8]) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_float_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        let fc = &mut self.fields[idx];
-        if idx >= 64 {
-            if fc.last_row == self.lifecycle.row_count() {
-                return;
-            }
-            fc.last_row = self.lifecycle.row_count();
-        }
-        if let Some(v) = parse_float_fast(value) {
-            fc.has_float = true;
-            fc.float_values.push((self.lifecycle.row_count(), v));
-        }
-    }
-
-    #[inline(always)]
-    pub fn append_f64_value_by_idx(&mut self, idx: usize, value: f64) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_f64_value_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        let fc = &mut self.fields[idx];
-        if idx >= 64 {
-            if fc.last_row == self.lifecycle.row_count() {
-                return;
-            }
-            fc.last_row = self.lifecycle.row_count();
-        }
-        fc.has_float = true;
-        fc.float_values.push((self.lifecycle.row_count(), value));
-    }
-
-    #[inline(always)]
-    pub fn append_bool_by_idx(&mut self, idx: usize, value: bool) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_bool_by_idx called outside of a row"
-        );
-        if check_dup_bits(self.lifecycle.written_bits_mut(), idx) {
-            return;
-        }
-        let fc = &mut self.fields[idx];
-        if idx >= 64 {
-            if fc.last_row == self.lifecycle.row_count() {
-                return;
-            }
-            fc.last_row = self.lifecycle.row_count();
-        }
-        fc.has_bool = true;
-        fc.bool_values.push((self.lifecycle.row_count(), value));
-    }
-
-    #[inline(always)]
-    pub fn append_null_by_idx(&mut self, idx: usize) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_null_by_idx called outside of a row"
-        );
-        // Nulls are represented by gaps -- no value record needed.
-        // But mark as written for duplicate-key detection.
-        let _ = check_dup_bits(self.lifecycle.written_bits_mut(), idx);
-
-        let fc = &mut self.fields[idx];
-        if idx >= 64 {
-            fc.last_row = self.lifecycle.row_count();
-        }
-    }
-
-    /// Store a zero-copy view of the raw unparsed line.
-    ///
-    /// Only has effect when the builder was created with line capture enabled.
-    /// The line must be a subslice of the buffer passed to `begin_batch`.
-    ///
-    /// First writer wins: if called multiple times in the same row, only the
-    /// first call has effect. This maintains the invariant that `line_views`
-    /// has exactly one entry per row when `line_capture` is enabled.
-    #[inline(always)]
-    pub fn append_line(&mut self, line: &[u8]) {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InRow,
-            "append_line called outside of a row"
-        );
-        if self.line_field_name.is_some() && !self.lifecycle.line_written_this_row() {
-            let line_view = if std::str::from_utf8(line).is_ok() {
-                let Ok(line_len) = u32::try_from(line.len()) else {
-                    return;
-                };
-                Some((self.offset_of(line), line_len))
-            } else {
-                let lossy = String::from_utf8_lossy(line);
-                let Ok(lossy_len) = u32::try_from(lossy.len()) else {
-                    return;
-                };
-                let Ok(decoded_offset) = u32::try_from(self.decoded_buf.len()) else {
-                    return;
-                };
-                let Some(combined_offset) = u32::try_from(self.buf.len())
-                    .ok()
-                    .and_then(|buf_len| buf_len.checked_add(decoded_offset))
-                else {
-                    return;
-                };
-                self.decoded_buf.extend_from_slice(lossy.as_bytes());
-                Some((combined_offset, lossy_len))
-            };
-            if let Some(line_view) = line_view {
-                self.line_views.push(line_view);
-            }
-            self.lifecycle.set_line_written();
-        }
-    }
-
-    /// Build a RecordBatch with zero-copy StringViewArrays.
-    ///
-    /// When no JSON escape sequences were decoded, the resulting RecordBatch
-    /// shares the input buffer via Bytes reference counting (zero-copy).
-    /// When decoded strings exist, the original and decoded buffers are exposed
-    /// as separate Arrow StringView blocks to avoid copying the whole input.
-    pub fn finish_batch(&mut self) -> Result<RecordBatch, ArrowError> {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InBatch,
-            "finish_batch called outside of a batch (call begin_batch first, and ensure all rows are closed with end_row)"
-        );
-        let num_rows = self.lifecycle.row_count() as usize;
-
-        // StringView offsets use original-buffer offsets for unescaped strings
-        // and offsets >= original_buf_len for decoded strings. Keep those as two
-        // Arrow blocks so decoding one field never copies the full input buffer.
-        let original_buf_len = u32::try_from(self.buf.len()).map_err(|_e| {
-            ArrowError::InvalidArgumentError("input buffer exceeds StringView offset range".into())
-        })?;
-        let arrow_buf = Buffer::from(self.buf.clone());
-        let decoded_arrow_buf = if self.decoded_buf.is_empty() {
-            None
-        } else {
-            Some(Buffer::from(self.decoded_buf.clone()))
-        };
-
-        let mut schema_fields: Vec<Field> = Vec::with_capacity(self.num_active);
-        let mut arrays: Vec<ArrayRef> = Vec::with_capacity(self.num_active);
-
-        // Detect duplicate output column names before building the schema.
-        let mut emitted_names = new_emitted_name_set();
-        if let Some(line_field_name) = self.line_field_name.as_ref() {
-            emitted_names.insert(line_field_name.clone());
-        }
-        let mut reserve_name = |name: &str| -> Result<(), ArrowError> {
-            if emitted_names.insert(name.to_string()) {
-                Ok(())
-            } else {
-                Err(ArrowError::InvalidArgumentError(format!(
-                    "duplicate output column name: {name}"
-                )))
-            }
-        };
-
-        for fc in &self.fields[..self.num_active] {
-            // Field names come from JSON keys (valid UTF-8 in well-formed input).
-            // Use from_utf8_lossy so that fuzz inputs with arbitrary bytes are
-            // handled gracefully instead of triggering undefined behaviour.
-            let name = String::from_utf8_lossy(&fc.name);
-            if self.line_field_name.as_deref() == Some(name.as_ref()) {
-                // Line capture owns this output column name for this batch.
-                // Keep scanner semantics as "line wins" when names collide.
-                continue;
-            }
-
-            // Emit a StructArray when the same field has multiple types in this
-            // batch. Single-type fields use the bare field name as a flat column.
-            let conflict = (fc.has_int as u8)
-                + (fc.has_float as u8)
-                + (fc.has_str as u8)
-                + (fc.has_bool as u8)
-                > 1;
-
-            if conflict {
-                let mut child_fields: Vec<Arc<Field>> = Vec::new();
-                let mut child_arrays: Vec<ArrayRef> = Vec::new();
-
-                if fc.has_int {
-                    let mut values = vec![0i64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.int_values {
-                        let row = row as usize;
-                        if row >= num_rows {
-                            continue;
-                        }
-                        values[row] = v;
-                        valid[row] = true;
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Int64Array::new(values.into(), Some(nulls));
-                    child_fields.push(Arc::new(Field::new("int", DataType::Int64, true)));
-                    child_arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_float {
-                    let mut values = vec![0.0f64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.float_values {
-                        let row = row as usize;
-                        if row >= num_rows {
-                            continue;
-                        }
-                        values[row] = v;
-                        valid[row] = true;
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Float64Array::new(values.into(), Some(nulls));
-                    child_fields.push(Arc::new(Field::new("float", DataType::Float64, true)));
-                    child_arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_str {
-                    let mut builder = StringViewBuilder::new();
-                    let original_block = builder.append_block(arrow_buf.clone());
-                    let decoded_block = decoded_arrow_buf
-                        .as_ref()
-                        .map(|buf| builder.append_block(buf.clone()));
-                    let mut vi = 0;
-                    for row in 0..num_rows as u32 {
-                        if vi < fc.str_views.len() && fc.str_views[vi].0 == row {
-                            let (_, offset, len) = fc.str_views[vi];
-                            append_string_view(
-                                &mut builder,
-                                original_block,
-                                decoded_block,
-                                original_buf_len,
-                                offset,
-                                len,
-                            )
-                            .expect("offset/len pre-validated by offset_of and UTF-8 check");
-                            vi += 1;
-                        } else {
-                            builder.append_null();
-                        }
-                    }
-                    child_fields.push(Arc::new(Field::new("str", DataType::Utf8View, true)));
-                    child_arrays.push(Arc::new(builder.finish()) as ArrayRef);
-                }
-
-                if fc.has_bool {
-                    let mut values = vec![false; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.bool_values {
-                        let row = row as usize;
-                        if row >= num_rows {
-                            continue;
-                        }
-                        values[row] = v;
-                        valid[row] = true;
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = BooleanArray::new(values.into(), Some(nulls));
-                    child_fields.push(Arc::new(Field::new("bool", DataType::Boolean, true)));
-                    child_arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                // Struct is non-null iff any child is non-null for that row.
-                let struct_validity: Vec<bool> = (0..num_rows)
-                    .map(|i| child_arrays.iter().any(|arr| !arr.is_null(i)))
-                    .collect();
-
-                let fields = Fields::from(child_fields);
-                let struct_arr = StructArray::new(
-                    fields.clone(),
-                    child_arrays,
-                    Some(NullBuffer::from(struct_validity)),
-                );
-
-                reserve_name(name.as_ref())?;
-                schema_fields.push(Field::new(name.as_ref(), DataType::Struct(fields), true));
-                arrays.push(Arc::new(struct_arr) as ArrayRef);
-            } else {
-                // Single-type field: bare flat column.
-                if fc.has_int {
-                    reserve_name(name.as_ref())?;
-                    let mut values = vec![0i64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.int_values {
-                        let row = row as usize;
-                        if row >= num_rows {
-                            continue;
-                        }
-                        values[row] = v;
-                        valid[row] = true;
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Int64Array::new(values.into(), Some(nulls));
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Int64, true));
-                    arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_float {
-                    reserve_name(name.as_ref())?;
-                    let mut values = vec![0.0f64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.float_values {
-                        let row = row as usize;
-                        if row >= num_rows {
-                            continue;
-                        }
-                        values[row] = v;
-                        valid[row] = true;
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Float64Array::new(values.into(), Some(nulls));
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Float64, true));
-                    arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_str {
-                    reserve_name(name.as_ref())?;
-                    let mut builder = StringViewBuilder::new();
-                    let original_block = builder.append_block(arrow_buf.clone());
-                    let decoded_block = decoded_arrow_buf
-                        .as_ref()
-                        .map(|buf| builder.append_block(buf.clone()));
-
-                    let mut vi = 0;
-                    for row in 0..num_rows as u32 {
-                        if vi < fc.str_views.len() && fc.str_views[vi].0 == row {
-                            let (_, offset, len) = fc.str_views[vi];
-                            append_string_view(
-                                &mut builder,
-                                original_block,
-                                decoded_block,
-                                original_buf_len,
-                                offset,
-                                len,
-                            )
-                            .expect("offset/len pre-validated by offset_of and UTF-8 check");
-                            vi += 1;
-                        } else {
-                            builder.append_null();
-                        }
-                    }
-
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Utf8View, true));
-                    arrays.push(Arc::new(builder.finish()) as ArrayRef);
-                }
-
-                if fc.has_bool {
-                    reserve_name(name.as_ref())?;
-                    let mut values = vec![false; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.bool_values {
-                        let row = row as usize;
-                        if row >= num_rows {
-                            continue;
-                        }
-                        values[row] = v;
-                        valid[row] = true;
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = BooleanArray::new(values.into(), Some(nulls));
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Boolean, true));
-                    arrays.push(Arc::new(array) as ArrayRef);
-                }
-            }
-        }
-
-        if self.line_field_name.is_some() {
-            // When line capture is enabled every row must have exactly one line entry.
-            // Check cardinality even when line_views is empty so that callers who
-            // never invoke append_line() get an explicit error rather than a batch
-            // silently missing the line column.
-            if self.line_views.len() != num_rows {
-                return Err(ArrowError::InvalidArgumentError(format!(
-                    "line_views cardinality mismatch: {} views for {} rows",
-                    self.line_views.len(),
-                    num_rows
-                )));
-            }
-            let mut builder = StringViewBuilder::new();
-            if num_rows > 0 {
-                let original_block = builder.append_block(arrow_buf);
-                let decoded_block = decoded_arrow_buf
-                    .as_ref()
-                    .map(|buf| builder.append_block(buf.clone()));
-                for row in 0..num_rows {
-                    let (offset, len) = self.line_views[row];
-                    append_string_view(
-                        &mut builder,
-                        original_block,
-                        decoded_block,
-                        original_buf_len,
-                        offset,
-                        len,
-                    )
-                    .expect("line view offset/len must be within buffer");
-                }
-            }
-            let line_field_name = self
-                .line_field_name
-                .as_deref()
-                .expect("line_field_name must be set when capture is enabled");
-            schema_fields.push(Field::new(line_field_name, DataType::Utf8View, true));
-            arrays.push(Arc::new(builder.finish()) as ArrayRef);
-        }
-
-        // Emit resource.attributes.* columns unconditionally (even for empty batches) so
-        // that the schema is identical regardless of row count. Arrow pipelines
-        // that concatenate or compare batches require a consistent schema; omitting
-        // these columns for num_rows == 0 would cause schema mismatch errors.
-        for (key, value) in &self.resource_attrs {
-            let col_name = Self::resource_col_name(key);
-            reserve_name(&col_name)?;
-            let mut builder = StringViewBuilder::new();
-            if num_rows > 0 {
-                let Ok(value_len) = u32::try_from(value.len()) else {
-                    return Err(ArrowError::InvalidArgumentError(format!(
-                        "resource attribute value too large for Utf8View: {key}"
-                    )));
-                };
-                let block = builder.append_block(Buffer::from(value.as_bytes().to_vec()));
-                for _ in 0..num_rows {
-                    builder
-                        .try_append_view(block, 0, value_len)
-                        .expect("resource attr constant view must be valid");
-                }
-            }
-            schema_fields.push(Field::new(col_name, DataType::Utf8View, true));
-            arrays.push(Arc::new(builder.finish()) as ArrayRef);
-        }
-
-        let schema = Arc::new(Schema::new(schema_fields));
-        let opts = RecordBatchOptions::new().with_row_count(Some(num_rows));
-        let result = RecordBatch::try_new_with_options(schema, arrays, &opts);
-        self.lifecycle.finish_batch();
-        result
-    }
-
-    /// Build a self-contained `RecordBatch` with detached `StringArray` columns.
-    ///
-    /// Uses the same scan data as `finish_batch` but produces `StringArray`
-    /// (contiguous offsets + data) instead of `StringViewArray` (views into
-    /// the input buffer). The input `Bytes` can be freed immediately after
-    /// this call.
-    ///
-    /// This is the optimal persistence path: zero-copy scan speed during
-    /// parsing, single bulk copy during finalization, and the resulting
-    /// `StringArray` compresses efficiently via IPC zstd.
-    pub fn finish_batch_detached(&mut self) -> Result<RecordBatch, ArrowError> {
-        debug_assert_eq!(
-            self.lifecycle.state(),
-            BuilderState::InBatch,
-            "finish_batch_detached called outside of a batch"
-        );
-        let num_rows = self.lifecycle.row_count() as usize;
-        let has_decoded = !self.decoded_buf.is_empty();
-
-        let mut schema_fields: Vec<Field> = Vec::with_capacity(self.num_active);
-        let mut arrays: Vec<ArrayRef> = Vec::with_capacity(self.num_active);
-
-        let mut emitted_names = new_emitted_name_set();
-        if let Some(line_field_name) = self.line_field_name.as_ref() {
-            emitted_names.insert(line_field_name.clone());
-        }
-        let mut reserve_name = |name: &str| -> Result<(), ArrowError> {
-            if emitted_names.insert(name.to_string()) {
-                Ok(())
-            } else {
-                Err(ArrowError::InvalidArgumentError(format!(
-                    "duplicate output column name: {name}"
-                )))
-            }
-        };
-
-        for fc in &self.fields[..self.num_active] {
-            let name = String::from_utf8_lossy(&fc.name);
-            if self.line_field_name.as_deref() == Some(name.as_ref()) {
-                // Line capture owns this output column name for this batch.
-                // Keep scanner semantics as "line wins" when names collide.
-                continue;
-            }
-            let conflict = (fc.has_int as u8)
-                + (fc.has_float as u8)
-                + (fc.has_str as u8)
-                + (fc.has_bool as u8)
-                > 1;
-
-            if conflict {
-                let mut child_fields: Vec<Arc<Field>> = Vec::new();
-                let mut child_arrays: Vec<ArrayRef> = Vec::new();
-
-                if fc.has_int {
-                    let mut values = vec![0i64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.int_values {
-                        let r = row as usize;
-                        if r < num_rows {
-                            values[r] = v;
-                            valid[r] = true;
-                        }
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Int64Array::new(values.into(), Some(nulls));
-                    child_fields.push(Arc::new(Field::new("int", DataType::Int64, true)));
-                    child_arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_float {
-                    let mut values = vec![0.0f64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.float_values {
-                        let r = row as usize;
-                        if r < num_rows {
-                            values[r] = v;
-                            valid[r] = true;
-                        }
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Float64Array::new(values.into(), Some(nulls));
-                    child_fields.push(Arc::new(Field::new("float", DataType::Float64, true)));
-                    child_arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_str {
-                    let total_bytes: usize = fc.str_views.iter().map(|&(_, _, l)| l as usize).sum();
-                    let mut builder =
-                        arrow::array::StringBuilder::with_capacity(num_rows, total_bytes);
-                    let mut vi = 0;
-                    for row in 0..num_rows as u32 {
-                        if vi < fc.str_views.len() && fc.str_views[vi].0 == row {
-                            let (_, offset, len) = fc.str_views[vi];
-                            if let Some(s) = self.read_str(offset, len, has_decoded) {
-                                builder.append_value(s);
-                            } else {
-                                builder.append_null();
-                            }
-                            vi += 1;
-                        } else {
-                            builder.append_null();
-                        }
-                    }
-                    child_fields.push(Arc::new(Field::new("str", DataType::Utf8, true)));
-                    child_arrays.push(Arc::new(builder.finish()) as ArrayRef);
-                }
-
-                if fc.has_bool {
-                    let mut values = vec![false; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.bool_values {
-                        let r = row as usize;
-                        if r < num_rows {
-                            values[r] = v;
-                            valid[r] = true;
-                        }
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = BooleanArray::new(values.into(), Some(nulls));
-                    child_fields.push(Arc::new(Field::new("bool", DataType::Boolean, true)));
-                    child_arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                let struct_validity: Vec<bool> = (0..num_rows)
-                    .map(|i| child_arrays.iter().any(|arr| !arr.is_null(i)))
-                    .collect();
-                let fields = Fields::from(child_fields);
-                let struct_arr = StructArray::new(
-                    fields.clone(),
-                    child_arrays,
-                    Some(NullBuffer::from(struct_validity)),
-                );
-                reserve_name(name.as_ref())?;
-                schema_fields.push(Field::new(name.as_ref(), DataType::Struct(fields), true));
-                arrays.push(Arc::new(struct_arr) as ArrayRef);
-            } else {
-                if fc.has_int {
-                    reserve_name(name.as_ref())?;
-                    let mut values = vec![0i64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.int_values {
-                        let r = row as usize;
-                        if r < num_rows {
-                            values[r] = v;
-                            valid[r] = true;
-                        }
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Int64Array::new(values.into(), Some(nulls));
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Int64, true));
-                    arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_float {
-                    reserve_name(name.as_ref())?;
-                    let mut values = vec![0.0f64; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.float_values {
-                        let r = row as usize;
-                        if r < num_rows {
-                            values[r] = v;
-                            valid[r] = true;
-                        }
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = Float64Array::new(values.into(), Some(nulls));
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Float64, true));
-                    arrays.push(Arc::new(array) as ArrayRef);
-                }
-
-                if fc.has_str {
-                    reserve_name(name.as_ref())?;
-                    let total_bytes: usize = fc.str_views.iter().map(|&(_, _, l)| l as usize).sum();
-                    let mut builder =
-                        arrow::array::StringBuilder::with_capacity(num_rows, total_bytes);
-                    let mut vi = 0;
-                    for row in 0..num_rows as u32 {
-                        if vi < fc.str_views.len() && fc.str_views[vi].0 == row {
-                            let (_, offset, len) = fc.str_views[vi];
-                            if let Some(s) = self.read_str(offset, len, has_decoded) {
-                                builder.append_value(s);
-                            } else {
-                                builder.append_null();
-                            }
-                            vi += 1;
-                        } else {
-                            builder.append_null();
-                        }
-                    }
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Utf8, true));
-                    arrays.push(Arc::new(builder.finish()) as ArrayRef);
-                }
-
-                if fc.has_bool {
-                    reserve_name(name.as_ref())?;
-                    let mut values = vec![false; num_rows];
-                    let mut valid = vec![false; num_rows];
-                    for &(row, v) in &fc.bool_values {
-                        let r = row as usize;
-                        if r < num_rows {
-                            values[r] = v;
-                            valid[r] = true;
-                        }
-                    }
-                    let nulls = NullBuffer::from(valid);
-                    let array = BooleanArray::new(values.into(), Some(nulls));
-                    schema_fields.push(Field::new(name.as_ref(), DataType::Boolean, true));
-                    arrays.push(Arc::new(array) as ArrayRef);
-                }
-            }
-        }
-
-        if self.line_field_name.is_some() {
-            // Same cardinality guard as the non-detached path: every row must
-            // have a line entry, including the zero-row case (0 == 0 passes).
-            if self.line_views.len() != num_rows {
-                return Err(ArrowError::InvalidArgumentError(format!(
-                    "line_views cardinality mismatch: {} views for {} rows",
-                    self.line_views.len(),
-                    num_rows
-                )));
-            }
-            let total_bytes: usize = self.line_views.iter().map(|&(_, l)| l as usize).sum();
-            let mut builder = arrow::array::StringBuilder::with_capacity(num_rows, total_bytes);
-            let has_decoded = !self.decoded_buf.is_empty();
-            for row in 0..num_rows {
-                let (offset, len) = self.line_views[row];
-                if let Some(s) = self.read_str(offset, len, has_decoded) {
-                    builder.append_value(s);
-                } else {
-                    builder.append_null();
-                }
-            }
-            let line_field_name = self
-                .line_field_name
-                .as_deref()
-                .expect("line_field_name must be set when capture is enabled");
-            schema_fields.push(Field::new(line_field_name, DataType::Utf8, true));
-            arrays.push(Arc::new(builder.finish()) as ArrayRef);
-        }
-
-        // Emit resource.attributes.* columns unconditionally (even for empty batches) so
-        // that the schema is identical regardless of row count. Arrow pipelines
-        // that concatenate or compare batches require a consistent schema; omitting
-        // these columns for num_rows == 0 would cause schema mismatch errors.
-        for (key, value) in &self.resource_attrs {
-            let col_name = Self::resource_col_name(key);
-            reserve_name(&col_name)?;
-            let mut builder =
-                arrow::array::StringBuilder::with_capacity(num_rows, num_rows * value.len());
-            for _ in 0..num_rows {
-                builder.append_value(value);
-            }
-            schema_fields.push(Field::new(col_name, DataType::Utf8, true));
-            arrays.push(Arc::new(builder.finish()) as ArrayRef);
-        }
-
-        let schema = Arc::new(Schema::new(schema_fields));
-        let opts = RecordBatchOptions::new().with_row_count(Some(num_rows));
-        let result = RecordBatch::try_new_with_options(schema, arrays, &opts);
-        self.lifecycle.finish_batch();
-        result
-    }
-
-    /// Read a string value from the buffer(s) by offset and length.
-    ///
-    /// Offsets `< buf.len()` read from the original input buffer.
-    /// Offsets `>= buf.len()` read from `decoded_buf` at `offset - buf.len()`.
-    fn read_str(&self, offset: u32, len: u32, has_decoded: bool) -> Option<&str> {
-        let start = offset as usize;
-        let end = start.checked_add(len as usize)?;
-        let buf_len = self.buf.len();
-        let bytes = if !has_decoded || start < buf_len {
-            self.buf.get(start..end)?
-        } else {
-            let dec_start = start.checked_sub(buf_len)?;
-            let dec_end = end.checked_sub(buf_len)?;
-            self.decoded_buf.get(dec_start..dec_end)?
-        };
-        std::str::from_utf8(bytes).ok()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
+//! Tests for the streaming zero-copy Arrow RecordBatch builder.
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use arrow::array::{Array, StructArray as ArrowStructArray};
+    use super::super::*;
+    use arrow::array::{
+        Array, Float64Array, Int64Array,
+        StructArray as ArrowStructArray,
+    };
+    use arrow::datatypes::DataType;
 
     #[test]
     fn test_basic_string_and_int() {
@@ -1979,14 +766,14 @@ mod tests {
             "3-way conflict must produce Struct, got {:?}",
             col.data_type()
         );
-        let sa = col.as_any().downcast_ref::<StructArray>().unwrap();
-        let names: Vec<&str> = sa.fields().iter().map(|f| f.name().as_str()).collect();
+        let sa = col.as_any().downcast_ref::<ArrowStructArray>().unwrap();
+        let names: Vec<&str> = sa.fields().iter().map(|f: &arrow::datatypes::FieldRef| f.name().as_str()).collect();
         assert!(names.contains(&"int"), "int child missing");
         assert!(names.contains(&"float"), "float child missing");
         assert!(names.contains(&"str"), "str child missing");
 
         // Row 0: int=42, float=null, str=null
-        let int_idx = sa.fields().iter().position(|f| f.name() == "int").unwrap();
+        let int_idx = sa.fields().iter().position(|f: &arrow::datatypes::FieldRef| f.name() == "int").unwrap();
         let int_col = sa
             .column(int_idx)
             .as_any()
@@ -2751,50 +1538,5 @@ mod tests {
             return a.value(row).to_string();
         }
         format!("<unknown type at row {row}>")
-    }
-}
-
-#[cfg(kani)]
-mod verification {
-    use super::*;
-
-    /// Prove begin_batch resets per-batch builder state while preserving the
-    /// reusable field storage.
-    ///
-    /// Unit tests cover the expensive `finish_batch_detached` / `RecordBatch`
-    /// integration path. Kani stays focused on the builder bookkeeping.
-    #[kani::proof]
-    #[kani::solver(kissat)]
-    fn verify_single_field_batch_created() {
-        let mut b = StreamingBuilder::new(Some("line".to_string()));
-        b.fields.push(FieldColumns::new(b"x"));
-        b.num_active = 1;
-        b.line_views.push((1, 2));
-        b.decoded_buf.extend_from_slice(b"decoded");
-        b.begin_batch(bytes::Bytes::from_static(b"test data pad"));
-        assert_eq!(b.num_active, 0);
-        assert_eq!(b.lifecycle.row_count(), 0);
-        assert!(b.line_views.is_empty());
-        assert!(b.decoded_buf.is_empty());
-        assert_eq!(b.lifecycle.state(), BuilderState::InBatch);
-        assert_eq!(b.fields.len(), 1);
-    }
-
-    /// Prove begin_row/end_row increments row_count exactly once per row.
-    #[kani::proof]
-    #[kani::unwind(5)]
-    #[kani::solver(kissat)]
-    fn verify_int_field_row_count_matches() {
-        let num_rows: u32 = kani::any();
-        kani::assume(num_rows <= 3);
-        let mut b = StreamingBuilder::new(None);
-        b.begin_batch(bytes::Bytes::from_static(b"pad"));
-        for _ in 0..num_rows {
-            b.begin_row();
-            b.end_row();
-        }
-        assert_eq!(b.lifecycle.row_count(), num_rows);
-        kani::cover!(num_rows == 0, "empty batch");
-        kani::cover!(num_rows > 0, "non-empty batch");
     }
 }
